validation 1.0 is ready

+++ src/main/java/pl/slawek/controller/CrudController.java
@@ -14,11 +14,13 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.ObjectError;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.validation.beanvalidation.LocalValidatorFactoryBean;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import pl.slawek.data.PlayerRepository;
 import pl.slawek.model.Player;
@@ -37,8 +39,9 @@ public class CrudController {
     }
     
 	@GetMapping("/manage")
-	public String managePlayers(Model model) {
+	public String managePlayers(Model model, @ModelAttribute("message") String message) {
 		model.addAttribute("player", new Player());
+		model.addAttribute("message", message);
 		return "manage";
 	}
 	
@@ -56,55 +59,47 @@ public class CrudController {
 			} 
         }
 	
-	
-	
-	/*@PostMapping("/save")
-	public String saveNewPlayer(@ModelAttribute Player player) {
-        try {
-        	playerRepository.save(player);
-        	return "redirect:/manage";
-        } catch(ConstraintViolationException e) {
-            Set<ConstraintViolation<?>> errors = e.getConstraintViolations();
-            errors.forEach(err -> System.err.println(
-                    err.getPropertyPath() + " " +
-                    err.getInvalidValue() + " " + 
-                    err.getMessage()));
-            return "manage";
-        }
-
-	}
-	
-	/*@PostMapping("/save")
-	public String saveNewPlayer(@Valid @ModelAttribute Player player, BindingResult result) {
-        if (result.hasErrors()) {
-            List<ObjectError> errors = result.getAllErrors();
-            errors.forEach(err -> System.out.println(err.getDefaultMessage()));
-        } else 
-        {
-        	playerRepository.save(player);
-        	return "redirect:/manage";
-        }
-		return "manage";
-	}*/
-	
 	@GetMapping("/find")
-	public String showPlayer(@RequestParam String nickname,Model model) {
-		if(nickname)
-		model.addAttribute("player",playerRepository.findFirstByNickName(nickname));
-		return "showplayer";
+	public String showPlayer(@RequestParam String nickname,Model model,RedirectAttributes redirectAttributes) {
+		Player player = playerRepository.findFirstByNickName(nickname);
+		if(player!=null)
+		{
+			model.addAttribute("player",player);
+			return "showplayer";
+		}else
+		{
+			redirectAttributes.addFlashAttribute("message", "Error - there is no such player");
+			return "redirect:/manage";
+		}
+		
 	}
 	
 	@PostMapping("/edit")
-	public String editPlayer(@RequestParam String nickname,Model model) {
-		model.addAttribute("player",playerRepository.findFirstByNickName(nickname));
-		return "editplayer";
+	public String editPlayer(@RequestParam String nickname,Model model, RedirectAttributes redirectAttributes) {
+		Player player = playerRepository.findFirstByNickName(nickname);
+		if(player!=null)
+		{
+			model.addAttribute("player",player);
+			return "editplayer";
+		}else
+		{
+			redirectAttributes.addFlashAttribute("message", "Error - there is no such player");
+			return "redirect:/manage";
+		}
 	}
 	
 	@PostMapping("/delete")
-	public String deletePlayer(@RequestParam String nickname,Model model) {
+	public String deletePlayer(@RequestParam String nickname,Model model, RedirectAttributes redirectAttributes) {
 		Player player = playerRepository.findFirstByNickName(nickname);
-		model.addAttribute("player",player);
-		playerRepository.delete(player);
-		return "showplayer";
+		if(player!=null)
+		{
+			model.addAttribute("player",player);
+			playerRepository.delete(player);
+			return "showplayer";
+		}else
+		{
+			redirectAttributes.addFlashAttribute("message", "Error - there is no such player");
+			return "redirect:/manage";
+		}
 	}
 }
